agents/rag-pipeline-agent/requirements.txt
llama-index>=0.10.0
llama-index-core>=0.10.0
click>=8.1.0
rich>=13.0.0
pyyaml>=6.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
